Fix unlimited comments being generated due to paging
This commit fixes the issue we run into, when the GET request for
listing the comments on a PR starts paging the response. Then the web
request doesn't get the latest comments and misses the one which was
added as a response to the test runs. This basically fixes the duplicate
comment check which prevents comments from being made multiple times for
a singe label.